+++ Menu-o-Matic/Menu-o-Matic.mm
@@ -85,7 +85,8 @@
       </li>
     </ul>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <hook NAME="MapStyle" background="#f9f9f8">
     <properties show_icon_for_attributes="true" edgeColorConfiguration="#808080ff,#ff0000ff,#0000ffff,#00ff00ff,#ff00ffff,#00ffffff,#7c0000ff,#00007cff,#007c00ff,#7c007cff,#007c7cff,#7c7c00ff" mapUsesOwnSaveOptions="true" save_modification_times="false" save_last_visited_node="default" show_note_icons="true" save_folding="never_save_folding" fit_to_viewport="false"/>
 
@@ -187,7 +188,8 @@
       To translate the description you have to define a translation for the key 'addons.${name}.description'.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="Menu-o-Matic for Freeplane!!&#xa;Menu-o-Matic is an AddOn for Freeplane that gives the user the possibility to create his/her own dialogs with the menu items he/she prefers.&#xa;It can also include scripts." ID="ID_1642042079"/>
 </node>
 <node TEXT="changes" POSITION="top_or_left" ID="ID_1349799607">
@@ -202,7 +204,8 @@
       Change log of this add-on: append one node for each noteworthy version and put the details for each version into a child node.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="v0.1.0" ID="ID_674021294">
 <node TEXT="initial release" ID="ID_17370257"/>
 </node>
@@ -264,9 +267,16 @@
 <node TEXT="Toolbar title can be modified in dialog" ID="ID_419612997"/>
 </node>
 <node TEXT="v0.1.9" ID="ID_1463454153">
-<node TEXT="Corrected button background color (bug in some L&amp;F)" ID="ID_1635777247"/>
+<node TEXT="Corrected button background color (bug in some L&amp;F)" ID="ID_1635777247">
+<node TEXT="In original Look&amp;Feels the buttons background weren&apos;t transparent" ID="ID_520607658"/>
+</node>
 <node TEXT="Added preference to (in)activate TitledBorders" ID="ID_1456132538"/>
-<node TEXT="Added popup menu to remove toolbars" ID="ID_1610581537"/>
+<node TEXT="Added popup menu to remove toolbars" ID="ID_1610581537">
+<node TEXT="Right clicking on toolbars" ID="ID_1636531086"/>
+</node>
+<node TEXT="Added powerButtons functionality" ID="ID_986228712">
+<node TEXT="Unifies multiple MenuActions nodes as a single &quot;powerButton&quot; that executes all actions at once" ID="ID_987728606"/>
+</node>
 </node>
 </node>
 <node TEXT="license" POSITION="top_or_left" ID="ID_492801974">
@@ -287,7 +297,8 @@
       The License text has to be entered as a child of the <i>'license'</i>&#xa0;node, either as plain text or as HTML.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="MIT License&#xa;&#xa;Copyright (c) 2021 Eduardo Frohlich.&#xa;&#xa;Permission is hereby granted, free of charge, to any person obtaining a copy&#xa;of this software and associated documentation files (the &quot;Software&quot;), to deal&#xa;in the Software without restriction, including without limitation the rights&#xa;to use, copy, modify, merge, publish, distribute, sublicense, and/or sell&#xa;copies of the Software, and to permit persons to whom the Software is&#xa;furnished to do so, subject to the following conditions:&#xa;&#xa;The above copyright notice and this permission notice shall be included in all&#xa;copies or substantial portions of the Software.&#xa;&#xa;THE SOFTWARE IS PROVIDED &quot;AS IS&quot;, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR&#xa;IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,&#xa;FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE&#xa;AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER&#xa;LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,&#xa;OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE&#xa;SOFTWARE." ID="ID_1379820487" VSHIFT_QUANTITY="-0.75 pt"/>
 </node>
 <node TEXT="preferences.xml" POSITION="top_or_left" ID="ID_809015286">
@@ -398,7 +409,8 @@
       </tr>
     </table>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <attribute_layout NAME_WIDTH="109.5 pt" VALUE_WIDTH="51.75 pt"/>
 <attribute NAME="runOnStartingFreeplane" VALUE="boolean"/>
 <attribute NAME="showMenuOMaticMenuLabels" VALUE="boolean"/>
@@ -425,7 +437,8 @@
       </li>
     </ul>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <attribute_layout NAME_WIDTH="166.5 pt" VALUE_WIDTH="37.5 pt"/>
 <attribute NAME="${name}_runOnStartingFreeplane" VALUE="true"/>
 <attribute NAME="${name}_showMenuOMaticMenuLabels" VALUE="true"/>
@@ -463,8 +476,9 @@
       </li>
     </ul>
   </body>
-</html></richcontent>
-<node TEXT="en" ID="ID_78974084">
+</html>
+</richcontent>
+<node TEXT="en" ID="ID_553754732">
 <attribute_layout NAME_WIDTH="234.74999 pt" VALUE_WIDTH="411.74999 pt"/>
 <attribute NAME="addons.${name}.lauchDialogFromNodes" VALUE="launch menu dialog directly"/>
 <attribute NAME="addons.${name}.1launchMenuOMaticDialog" VALUE="launch Menu-o-Matic dialog"/>
@@ -487,6 +501,7 @@
 <attribute NAME="addons.${name}.listCustomDialogs" VALUE="list opened menu dialogs"/>
 <attribute NAME="addons.${name}.unpackCustomMenu" VALUE="unpack menu"/>
 <attribute NAME="addons.${name}.launchDialogsFromBranch" VALUE="launch all dialog packages from branch"/>
+<attribute NAME="addons.${name}.markAsPowerButton" VALUE="Mark / unmark node as powerButton"/>
 </node>
 </node>
 <node TEXT="deinstall" POSITION="top_or_left" ID="ID_1351206087">
@@ -501,7 +516,8 @@
       List of files and/or directories to remove on uninstall
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <attribute_layout NAME_WIDTH="42.75 pt" VALUE_WIDTH="316.49999 pt"/>
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}.script.xml"/>
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}/scripts/launchCustomDialog.groovy"/>
@@ -532,6 +548,7 @@
 <attribute NAME="delete" VALUE="${installationbase}/icons/Menu-o-Matic/MoMCloseTab.svg"/>
 <attribute NAME="delete" VALUE="${installationbase}/icons/Menu-o-Matic/MoMCloseToolbar.svg"/>
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}/scripts/changePackParameters.groovy"/>
+<attribute NAME="delete" VALUE="${installationbase}/addons/${name}/scripts/markAsPowerButton.groovy"/>
 </node>
 <node TEXT="scripts" POSITION="bottom_or_right" ID="ID_1105183990">
 <edge COLOR="#007c00"/>
@@ -650,7 +667,8 @@
       &#xa0;&#xa0;- In any case set execute_scripts_without_asking to true unless you want to annoy users.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="getMenuCommand.groovy" ID="ID_1857102818">
 <attribute_layout NAME_WIDTH="204.74999 pt" VALUE_WIDTH="184.49999 pt"/>
 <attribute NAME="menuTitleKey" VALUE="addons.${name}.getMenuCommand"/>
@@ -688,6 +706,7 @@
 <attribute NAME="execute_scripts_without_network_restriction" VALUE="false"/>
 </node>
 <node TEXT="changePackParameters.groovy" ID="ID_439163238">
+<attribute_layout NAME_WIDTH="188.24999 pt" VALUE_WIDTH="170.24999 pt"/>
 <attribute NAME="menuTitleKey" VALUE="addons.${name}.changePackParameters"/>
 <attribute NAME="menuLocation" VALUE="${addonsMenu}/addons.${name}/design"/>
 <attribute NAME="executionMode" VALUE="on_single_node"/>
@@ -761,6 +780,18 @@
 <node TEXT="insertMenuOMaticPackage.groovy" ID="ID_1045039173">
 <attribute_layout NAME_WIDTH="204.74999 pt" VALUE_WIDTH="194.24999 pt"/>
 <attribute NAME="menuTitleKey" VALUE="addons.${name}.insertMenuOMaticPackage"/>
+<attribute NAME="menuLocation" VALUE="${addonsMenu}/addons.${name}/design"/>
+<attribute NAME="executionMode" VALUE="on_single_node"/>
+<attribute NAME="keyboardShortcut" VALUE=""/>
+<attribute NAME="execute_scripts_without_asking" VALUE="true"/>
+<attribute NAME="execute_scripts_without_file_restriction" VALUE="true"/>
+<attribute NAME="execute_scripts_without_write_restriction" VALUE="false"/>
+<attribute NAME="execute_scripts_without_exec_restriction" VALUE="false"/>
+<attribute NAME="execute_scripts_without_network_restriction" VALUE="false"/>
+</node>
+<node TEXT="markAsPowerButton.groovy" ID="ID_1919407952">
+<attribute_layout NAME_WIDTH="188.24999 pt" VALUE_WIDTH="159.75 pt"/>
+<attribute NAME="menuTitleKey" VALUE="addons.${name}.markAsPowerButton"/>
 <attribute NAME="menuLocation" VALUE="${addonsMenu}/addons.${name}/design"/>
 <attribute NAME="executionMode" VALUE="on_single_node"/>
 <attribute NAME="keyboardShortcut" VALUE=""/>
@@ -846,7 +877,8 @@
       &#xa0;- The files will be processed in the sequence as seen in the map.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="Menu-o-Matic.jar" ID="ID_257757258"/>
 </node>
 <node TEXT="zips" POSITION="bottom_or_right" ID="ID_532152967">
@@ -885,7 +917,8 @@
       &#xa0;- The files will be processed in the sequence as seen in the map.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="doc" ID="ID_108012296"/>
 <node TEXT="templates" ID="ID_1212139449"/>
 <node TEXT="scripts" ID="ID_162715608"/>
@@ -923,7 +956,8 @@
       Images can be added automatically by releaseAddOn.groovy or must be uploaded into the map via the script <i>Tools-&gt;Scripts-&gt;Insert Binary</i>&#xa0;since they have to be (base64) encoded as simple strings.
     </p>
   </body>
-</html></richcontent>
+</html>
+</richcontent>
 <node TEXT="menuOMatic-screenshot-1.png" ID="ID_828445446"/>
 <node TEXT="menuOMatic-icon.png" STYLE_REF="freshNew" ID="ID_787195450">
 <richcontent TYPE="DETAILS" CONTENT-TYPE="plain/markdown"/>
@@ -942,11 +976,12 @@
   <body>
     Direct links to menu commands
   </body>
-</html></richcontent>
-<node TEXT="Build add-on" ID="ID_1006867388" LINK="menuitem:_addons.devtools.checkAddOn_on_single_node"/>
-<node TEXT="Package add-on for publication" ID="ID_1119353524" LINK="menuitem:_addons.devtools.releaseAddOn_on_single_node"/>
-<node TEXT="Export Translations" ID="ID_1088411781" LINK="menuitem:_addons.devtools.exportTranslations_on_single_node"/>
-<node TEXT="Import Translations" ID="ID_1039784888" LINK="menuitem:_addons.devtools.importTranslations_on_single_node"/>
+</html>
+</richcontent>
+<node TEXT="Build add-on[translate me]" ID="ID_1784712276" LINK="menuitem:_addons.devtools.checkAddOn_on_single_node"/>
+<node TEXT="Package add-on for publication[translate me]" ID="ID_1042673242" LINK="menuitem:_addons.devtools.releaseAddOn_on_single_node"/>
+<node TEXT="Export Translations[translate me]" ID="ID_1190923889" LINK="menuitem:_addons.devtools.exportTranslations_on_single_node"/>
+<node TEXT="Import Translations[translate me]" ID="ID_1711584072" LINK="menuitem:_addons.devtools.importTranslations_on_single_node"/>
 </node>
 </node>
 </map>